Check OpenNIC dig size before overwritting zone file
Checking for a size of 100 Bytes. This is based on existing valid zones.
Smallest is just above that. An empty one will be much less.

This ensures we don't overwrite a zone file in case T1 returns wrong
data.

Also restart pdns in the end for changes to take effect